Change to float again

+++ src/trip/index.js
@@ -74,7 +74,7 @@ function initSchema(config) {
       tagline: String
       description: String
       shortDescription: String
-      distance: Int
+      distance: Float
       lat: Float
       lng: Float
       season: String
